ControlledOperation._apply_unitary_
