+++ cirq/ops/controlled_operation_test.py
@@ -16,6 +16,8 @@ import pytest
 import numpy as np
 import cirq
 from cirq import protocols
+from cirq.ops.controlled_test_util import GateUsingWorkspaceForApplyUnitary, \
+    GateAllocatingNewSpaceForResult
 
 
 def test_controlled_operation_init():
@@ -141,7 +143,9 @@ def test_non_diagrammable_subop():
     cirq.SWAP(cirq.NamedQubit('q1'), cirq.NamedQubit('q2')),
     cirq.CCZ(cirq.NamedQubit('q1'), cirq.NamedQubit('q2'),
              cirq.NamedQubit('q3')),
-    cirq.ControlledGate(cirq.ControlledGate(cirq.CCZ))(*cirq.LineQubit.range(5))
+    cirq.ControlledGate(cirq.ControlledGate(cirq.CCZ))(*cirq.LineQubit.range(5)),
+    GateUsingWorkspaceForApplyUnitary()(cirq.NamedQubit('q1')),
+    GateAllocatingNewSpaceForResult()(cirq.NamedQubit('q1')),
 ])
 def test_controlled_operation_is_consistent(gate: cirq.GateOperation):
     cb = cirq.NamedQubit('ctr')
